+++ packages/pywavelets/test_pywt.py
@@ -1,7 +1,7 @@
 from pyodide_build.testing import run_in_pyodide
 
 
-@run_in_pyodide(packages=["pywavelets"])
+@run_in_pyodide(packages=["pywavelets"], driver_timeout=30)
 def test_pywt():
     import pywt
     import numpy as np