+++ docs/project/changelog.md
@@ -47,7 +47,7 @@ substitutions:
   Made a `PyProxy` of a Python generator into a Javascript generator:
   `proxy.next(val)` is translated to `gen.send(val)`.
   [#1180](https://github.com/pyodide/pyodide/pull/1180)
-- Updated `PyProxy` so that if the wrapped Python object supports `__getitem__`
+- {{ API }} Updated `PyProxy` so that if the wrapped Python object supports `__getitem__`
   access, then the wrapper has `get`, `set`, `has`, and `delete` methods which do
   `obj[key]`, `obj[key] = val`, `key in obj` and `del obj[key]` respectively.
   [#1175](https://github.com/pyodide/pyodide/pull/1175)
@@ -56,6 +56,15 @@ substitutions:
 - {{ API }} Added `PyProxy.getBuffer` API to allow direct access to Python
   buffers as Javascript TypedArrays.
   [1215](https://github.com/pyodide/pyodide/pull/1215)
+- {{ API }} The innermost level of a buffer converted to Javascript used to be a
+  TypedArray if the buffer was contiguous and otherwise an Array. Now the
+  innermost level will be a TypedArray unless the buffer format code is a '?' in
+  which case it will be an Array of booleans, or if the format code is a "s" in
+  which case the innermost level will be converted to a string.
+  [1376](https://github.com/pyodide/pyodide/pull/1376)
+- {{ Enhancement }} Javascript `BigInt`s are converted into Python `int` and
+  Python `int`s larger than 2^53 are converted into `BigInt`.
+  [1407](https://github.com/pyodide/pyodide/pull/1407)
 
 ### Fixed
 - {{ Fix }} getattr and dir on JsProxy now report consistent results and include all